refactor: improve the user constructor
기존 빌더 패턴은 의도를 명확히 드러내지 않습니다.
정적 팩토리 메서드를 통해 명확한 의도를 나타내고
내부에 빌더 패턴을 사용함으로써 가독성을 높였습니다.
