--- .gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,0 @@
-#Wed Sep 18 18:25:18 KST 2024
-gradle.version=8.10.1
