+++ src/main/java/lems/cowshed/config/jwt/JwtFilter.java
@@ -54,13 +54,7 @@ public class JwtFilter extends OncePerRequestFilter {
         String username = jwtUtil.getUsername(token);
         String role = jwtUtil.getRole(token);
 
-        User user = User.builder()
-                .id(userId)
-                .email(email)
-                .username(username)
-                .password("tempPass")
-                .role(role)
-                .build();
+        User user = User.createUserDetails(userId, username, "tempPass", role, email);
 
         //UserDetails에 회원 정보 객체 담기
         CustomUserDetails customUserDetails = new CustomUserDetails(user);
